fix configurations

+++ src/main/resources/schema.sql
@@ -3,7 +3,7 @@ CREATE TABLE transactions (
     tid VARCHAR(255) NOT NULL UNIQUE,
     from_user VARCHAR(255),
     to_user VARCHAR(255),
-    value_curr DECIMAL(10, 2),
+    value_curr FLOAT,
     type VARCHAR(255),
     create_time BIGINT
 );
